Include cstring required for GCC 4.3
GCC 4.3 has tightened up header dependencies so that
`strcmp' is no longer declared if <cstdlib> is included
(which it was in prior versions). This causes the changes
in 5e512e65704728b52e28448b4c18c70815a5b4b0 to fail to
build with 4.3.

This patch includes <cstring> in the files that are failing.